feat: ensure rust closures live long enough
This uses `.forget` to ensure the Closure will not have been freed
before being used on the JavaScript side. My current understanding is
that this will leak the memory but will not be an issue because it is
only happening in a few places, and the memory should be freed when the
service worker is unloaded. If that is not the case this will need to be
amended.

Additionally the worker is now packaged as an es module. This currently
works on most browsers, however there is an exception for Firefox. If
using Firefox then at least version 111 will need to be used and a flag
will need to be enabled to allow for workers to be a module.